feat(6502): native C engine compiled to a real 6502 via llvm-mos (~1830 on-chip)
The hand-written 6502 search tops out around ~1072 on-chip even at depth 6
(its root heuristics and depth/timeout reality cap it). Rather than grind
that up, compile the already-strong native C engine DOWN to the 6502 with
llvm-mos (LLVM backend, zero-page register allocation -- not cc65's
software-stack tax, which was 25-30x too slow). The same C the host oracle
runs now produces 6502 machine code.

What landed (all under tools/llvmmos_bench/, host-guarded edits in native/):
- The full engine (board + movegen + full eval + alpha-beta + TT +
  quiescence + killers + null-move + LMR) LINKS and FITS in 64K via the
  mos-sim target. RAM diet (guarded under __CC65__||__mos__ so the host
  build + all host gates are byte-identical): MAX_PLY 7, shared quiescence
  move buffer (g_filt indexed by quiescence depth), history heuristic off
  (16K table won't fit; TT move + MVV-LVA + killers + null-move + LMR stay),
  TT_BITS 8, MAX_PATH 64. board.h hash_t pinned to uint32_t so the splitmix
  zobrist computes identical keys on the 64-bit host and the 32-bit 6502.
- engine6502.c: a FEN->bestmove ABI driver (g_fen/g_depth/g_go ->
  bench_bestmove -> g_from/g_to/g_promo). validate.c drives the .sim image
  inside the cycle-exact fast6502 core and compares move-for-move to the
  host. engine6502_cli.c is a cref-compatible CLI over the real image so
  native_vs_stockfish (NATIVE_CREF=...) measures the ACTUAL binary vs SF.
- build_engine6502.sh rebuilds + revalidates; NOTES.md documents the ABI,
  the RAM diet, and the one-time llvm-mos toolchain/SDK setup.

VALIDATION: the 6502 image's best move is BIT-IDENTICAL to the host engine
in the same config -- depth 4 60/60, depth 6 6/6. STRENGTH (that exact
config, depth 6, native_vs_stockfish): vs SF-1500 67/80 = 83.8% (~1785),
vs SF-1700 58.5/80 = 73.1% (~1874) -> ~1830, comfortably above 1800. At
1MHz it's ~200s/move (correspondence clock; irrelevant to move quality).

Host gates unchanged: eval 22157/22157 bit-exact, PERFT EXACT.

+++ native/board.h
@@ -14,11 +14,16 @@
 
 #include <stdint.h>
 
-/* Zobrist/repetition key type. cc65 (6502) has no 64-bit integers, so we use
- * `unsigned long`, which is >=32 bits on every target: 32-bit on cc65 (the hash
- * narrows to 32 bits there -- fine, it only keys the TT/repetition table) and
- * 64-bit on the LP64 host. Use hash_t everywhere a zobrist key is stored. */
-typedef unsigned long hash_t;
+/* Zobrist/repetition key type. FIXED at exactly 32 bits (uint32_t) on EVERY
+ * target. This matters for bit-exact move agreement between the host and the
+ * 6502 (llvm-mos) port: the splitmix zobrist mixer folds high bits down with
+ * shifts, so a 64-bit `unsigned long` host and a 32-bit 6502 would compute
+ * DIFFERENT keys -> different TT collisions and repetition hits -> occasionally
+ * different best moves. Pinning to uint32_t makes the keys identical on both, so
+ * the 6502 reproduces the host search exactly. 32 bits is ample to key the TT
+ * (<=64K entries) and the repetition table. (perft counts at the depths we verify
+ * stay < 2^32, so the `hash_t perft()` return is unaffected.) */
+typedef uint32_t hash_t;
 
 #define WHITE_FLAG 0x80
 #define PT(p)       ((p) & 7)          /* piece type 1..6, 0 if empty */
